+++ jobbergate-agent/CHANGELOG.md
@@ -4,14 +4,20 @@ This file keeps track of all notable changes to jobbergate-core
 
 ## Unreleased
 
+- Capture request errors with Slurm API in Sentry notifications [PENG-2000]
+- Capture request errors with Jobbergate API in Sentry notifications [PENG-2000]
 
 ## 4.2.0a8 -- 2024-01-05
 ## 4.2.0a7 -- 2023-12-13
+
 ## 4.2.0a6 -- 2023-12-12
+
 ## 4.2.0a5 -- 2023-12-12
+
 - Added support for Python 3.12
 
 ## 4.2.0a4 -- 2023-12-11
+
 ## 4.2.0a3 -- 2023-11-30
 
 ## 4.2.0a2 -- 2023-11-29